chapter_002/src/main/java/ru/job4j/pseudo/Triangle.java
package ru.job4j.pseudo;
/**
 * Triangle
 * @author Sachenkov Maxim (oqiwhite1996@gmail.com)
 * @version $Id$
 * @since 0.1
 */
public class Triangle implements Shape {
    /**
     * Переопределенный метод draw
     * @return возвращает в виде строковых значений треугольник.
     */
    @Override
    public String draw() {
        StringBuilder pic = new StringBuilder();
        pic.append("   +   ");
        pic.append("  + +  ");
        pic.append(" +   + ");
        pic.append("+++++++");
        return pic.toString();
    }
}
